refactor: 게스트 초대 취소 로직 검증 이동
